moving relation status scale bar
